fix(ci): stop the metrics test forcing a workspace rebuild (fixes #12381)
`make nextest` ran the query-metrics test as its own cargo invocation:

    cargo nextest run -p runtime --test query_metrics

Cargo resolves features per package selection, and `crates/runtime` declares
no `default` feature, so that selection gave `runtime` 6 features (only what
the connector dev-dep cycles pull in) where the workspace line gives it 35.
Different features means different fingerprints, so `runtime` and hundreds of
its dependencies recompiled from scratch to produce one test binary holding
one 0.6s test — 42 minutes of `Build and Test`.

Target filters do not affect the resolve, so the test target moves onto the
existing workspace line and costs a codegen plus a link instead. Measured on
a cold tree: the lib-only line takes 43m38s, and adding `--test metrics` to
it then takes 8s.

Fold the three metrics test binaries into one while moving them, because only
`query_metrics` was ever named in the Makefile: `memory_metrics` and
`query_failure_err_code` were compiled by `make lint-rust` and run by nothing,
so their assertions never gated a merge. Naming the binary rather than each
file also means the next metrics test needs no CI wiring, which is how those
two came to be dead. The install helper they each duplicated, which has to
keep mirroring what `spiced` builds for `--metrics`, now has one copy.

One provider serves the whole binary, installed through a `LazyLock` so the
order holds under both harnesses: nextest gives each test its own process,
`cargo test` gives them threads. That makes the registry shared state, so the
pool-refusal test compares a before/after delta rather than asserting an
absolute count — it previously passed only because the sibling binary's
failing query happened to map to a different `err_code`.

The `-p cayenne` line has the same defect (131 packages resolve different
features than the workspace line, including tokio, bytes and arrow). Folding
it in changes what its test binaries compile against, so it is left with a
pointer to #12504.
